Add additional guard for $DoneCount against illogical values
It's like a race condition, caused by a very unfortunate timing of events, when additional files are written to the processed directory by an external process, between the steps of counting items. This is very rare, I've seen this only happen once in thousands of executions of this script (tested via automatic execution in a loop).
It should be theoretically possible to reproduce such a specific case, but not without jumping through hundreds of hoops, so I can't really be bothered right now.